+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.5.0 (Unreleased)
+
+- ``find_path`` has been renamed ``find_polygon_path``
+- A new ``find_path`` function has been added that does both polygon path finding and stringpulling in one.
+- ``OxidizedNavigationPlugin`` now takes a settings parameter containing ``NavMeshSettings`` 
+
+## 0.4.0 (2023-04-11)
+
+- Nav-Mesh generation now correctly reacts to removing the ``NavMeshAffector`` component.
+- ``NavMeshAffector`` is now an empty type, the data has been moved to a resource to enable the above change.
+
 ## 0.3.0 (2023-03-07)
 
 - Updated to Bevy 0.10